% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npawa.R
\name{npawa}
\alias{npawa}
\title{Raising a number to the power of nine
Returns a result of a number raised to power nine}
\usage{
npawa(x)
}
\arguments{
\item{x}{variable of class 'numeric'}
}
\value{
numeric
}
\description{
Raising a number to the power of nine
Returns a result of a number raised to power nine
}
